backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyMuPDF==1.24.0
python-docx==1.1.0
spacy==3.7.5
scikit-learn==1.5.0
motor==3.5.0
pydantic==2.7.0
google-generativeai==0.8.3
python-dotenv==1.0.1
